Optimize ETC1 selector palette matching
Right now when compression level isn't 0, we take each 4x4 block and try
to replace the selector indices with one of the 64 entries in the
palette. If we find a resulting selector block that has a low enough
error, we use that instead.

The way the code is implemented right now, we decode each block from
scratch for each entry, which means we decode each pixel 64 times.

However, each pixel's decoded result only depends on the endpoints (that
are static) and on the selector in that pixel (which has only 4 possible
values), so for each pixel we need at most 4 decodes.

This change switches the code around so that we pre-decode all 4
variants of the entire block ahead of time, and then just index into
this when evaluating matches.

This results in bit-identical images since we look at the exact same
data, but the encoding is faster - on a 2Kx2K PNG on a Core i7-8700K,
this change reduces the encoding time including a full mip chain from
8.5 seconds to 7.3 seconds.