Optimize color_distance function
This change optimizes the perceptual color distance function in a few
steps:

1. Instead of converting each color to YCrCb, and then taking the
deltas, we compute the deltas in RGB and then convert deltas to YCrCb
space. This is using the fact that (R1 - Y1) - (R2 - Y2) = (R1 - R2) -
(Y1 - Y2), and Y delta can be computed from R/G/B deltas using the same
coefficients.

2. Instead of computing color distances one at a time, we introduce an
SSE2-optimized function that computes the distance from one color to a
block of 4. It also returns the index of the minimum error since that's
used by almost every caller. All performance-sensitive locations where
this is feasible are converted to color_distance4.

3. In find_optimal_selector_clusters_for_each_block, a lot of time was
taken computing color distances, and they didn't form a neat groups of
4. However, because we only have 4 colors to compare against, and 16
colors to choose from, we have to compute a grand total of 64 unique
color deltas; the cluster count, however, is typically ~200 and we
computed 16 deltas for each cluster - so it's cheaper to precompute all
64 deltas and just add them up.

Together these optimizations reduce the time to compress a 2Kx2K color
image from 6.6 sec (post #105) to 4.7 sec, including image load and mipmap
generation time.

Note that the results aren't bit-exact due to the small floating point
drift in perceptually correct color distance, but PSNR is the same as
before.