Remove contention point from find_optimal_selector_clusters_for_each_block
Before this change we filled block->cluster mapping in parallel but also
tried to aggregate list of blocks for each cluster in the same job.

In a multithreaded run on a Ryzen 5900 system this resulted in
significant contention over the mutex; moving the distribution to run
after all threads finished (serially) brought total compression time
down from 6.5 to 5.5 seconds.