Rewrite job_pool to make it possible to wait for a series of jobs
Now add_job accepts an optional token; when wait_for_all gets the same
token, it only steals the job with that token and returns when all jobs
submitted with the token have finished. This allows us to use a single
job pool for many concurrent textures that are being compressed, which
results in much better scalability vs using wait_for_all() without
tokens.

The token is simply the number of outstanding jobs; the lifetime of that
token is managed by the caller, so wait_for_all needs to be called
before the token dies.